+++ formula-test/src/main/java/com/instacart/formula/test/TestRuntimeExtensions.kt
@@ -2,37 +2,27 @@ package com.instacart.formula.test
 
 import com.instacart.formula.IFormula
 import com.instacart.formula.Stream
-import io.reactivex.rxjava3.core.Observable
 
 /**
  * An extension function to create a [TestFormulaObserver] for a [IFormula] instance.
  *
- * @param input Input passed to [IFormula].
+ * Note: Formula won't start until you pass it an [input][TestFormulaObserver.input].
  */
-fun <Input : Any, Output : Any, F: IFormula<Input, Output>> F.test(
-    input: Input
-): TestFormulaObserver<Input, Output, F> {
-    return test(Observable.just(input))
+fun <Input : Any, Output : Any, F: IFormula<Input, Output>> F.test(): TestFormulaObserver<Input, Output, F> {
+    return TestFormulaObserver(this)
 }
-
 
 /**
  * An extension function to create a [TestFormulaObserver] for a [IFormula] instance.
  *
- * @param input A stream of inputs passed to [IFormula].
+ * @param initialInput Input passed to [IFormula].
  */
 fun <Input : Any, Output : Any, F: IFormula<Input, Output>> F.test(
-    input: Observable<Input>
+    initialInput: Input
 ): TestFormulaObserver<Input, Output, F> {
-    return TestFormulaObserver(
-        input = input,
-        formula = this
-    )
+    return test().apply {
+        input(initialInput)
+    }
 }
-
-/**
- * An extension function to create a [TestFormulaObserver] for a [IFormula] instance.
- */
-fun <Output : Any, F: IFormula<Unit, Output>> F.test() = test(Unit)
 
 fun <Message> Stream<Message>.test() = TestStreamObserver(this)