+++ formula/src/test/java/com/instacart/formula/ManyEmissionStreamTest.kt
@@ -13,7 +13,7 @@ class ManyEmissionStreamTest {
 
     @Test fun `all increment events go through`() {
         TestFormula()
-            .test()
+            .test(Unit)
             .apply {
                 assertThat(values()).containsExactly(EMISSION_COUNT).inOrder()
             }